@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:google_maps_place_picker/google_maps_place_picker.dart';
 import 'package:morea/Widgets/standart/buttons.dart';
 import 'package:morea/Widgets/standart/moreaTextStyle.dart';
+import 'package:morea/morea_strings.dart';
 import 'package:morea/morealayout.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:morea/services/utilities/moreaInputValidator.dart';
 
 class ChangeAbtreten extends StatefulWidget {
@@ -20,10 +19,10 @@ class ChangeAbtreten extends StatefulWidget {
 }
 
 class _ChangeAbtretenState extends State<ChangeAbtreten> {
-  String ortAbtreten;
-  String zeitAbtreten;
-  String urlMapAbtreten;
-  String nameMapAbtreten;
+  late String ortAbtreten;
+  late String zeitAbtreten;
+  late String urlMapAbtreten;
+  late String nameMapAbtreten;
   final _formKey = GlobalKey<FormState>();
   TextEditingController ortAbtretenController = TextEditingController();
 
@@ -146,7 +145,7 @@ class _ChangeAbtretenState extends State<ChangeAbtreten> {
                                       BorderSide(color: MoreaColors.violett)),
                             ),
                             validator: (value) {
-                              if (value.isEmpty) {
+                              if (value!.isEmpty) {
                                 return 'Bitte nicht leer lassen';
                               } else {
                                 return null;
@@ -163,7 +162,110 @@ class _ChangeAbtretenState extends State<ChangeAbtreten> {
                         ),
                         Padding(
                             padding: const EdgeInsets.only(top: 10.0),
-                            child: moreaRaisedButton(this.nameMapAbtreten, () {
+                            child: Column(
+                              children: [
+                                TextButton(
+                                  child: Container(
+                                    constraints: BoxConstraints(
+                                        minWidth: viewportConstraints.maxWidth,
+                                        maxWidth: viewportConstraints.maxWidth),
+                                    child: Padding(
+                                      padding: const EdgeInsets.all(8.0),
+                                      child: Text('Lokal',
+                                        style: MoreaTextStyle.textField,
+                                      ),
+                                    ),
+                                  ),
+                                  style: ButtonStyle(
+                                      shape: MaterialStateProperty.all<OutlinedBorder>(
+                                          RoundedRectangleBorder(
+                                              borderRadius: BorderRadius.circular(5),
+                                              side: BorderSide(color: Colors.black45))),
+                                      foregroundColor: MaterialStateProperty.all<Color>(
+                                          Colors.black),
+                                      padding: MaterialStateProperty.all<EdgeInsets>(
+                                          EdgeInsets.only(
+                                              top: 10, bottom: 10, left: 5, right: 5)),
+                                      overlayColor: MaterialStateProperty.resolveWith(
+                                              (Set<MaterialState> states) {
+                                            if (states.contains(MaterialState.focused))
+                                              return MoreaColors.violett;
+                                            return null;
+                                          })),
+                                  onPressed: () {
+                                    this.urlMapAbtreten = urlGoogleMapsLokal;
+                                    this.nameMapAbtreten = 'Lokal';
+                                  },
+                                ),
+                                TextButton(
+                                  child: Container(
+                                    constraints: BoxConstraints(
+                                        minWidth: viewportConstraints.maxWidth,
+                                        maxWidth: viewportConstraints.maxWidth),
+                                    child: Padding(
+                                      padding: const EdgeInsets.all(8.0),
+                                      child: Text('Waldrand',
+                                        style: MoreaTextStyle.textField,
+                                      ),
+                                    ),
+                                  ),
+                                  style: ButtonStyle(
+                                      shape: MaterialStateProperty.all<OutlinedBorder>(
+                                          RoundedRectangleBorder(
+                                              borderRadius: BorderRadius.circular(5),
+                                              side: BorderSide(color: Colors.black45))),
+                                      foregroundColor: MaterialStateProperty.all<Color>(
+                                          Colors.black),
+                                      padding: MaterialStateProperty.all<EdgeInsets>(
+                                          EdgeInsets.only(
+                                              top: 10, bottom: 10, left: 5, right: 5)),
+                                      overlayColor: MaterialStateProperty.resolveWith(
+                                              (Set<MaterialState> states) {
+                                            if (states.contains(MaterialState.focused))
+                                              return MoreaColors.violett;
+                                            return null;
+                                          })),
+                                  onPressed: () {
+                                    this.urlMapAbtreten = urlGoogleMapsWaldrand;
+                                    this.nameMapAbtreten = 'Waldrand';
+                                  },
+                                ),
+                                TextButton(
+                                  child: Container(
+                                    constraints: BoxConstraints(
+                                        minWidth: viewportConstraints.maxWidth,
+                                        maxWidth: viewportConstraints.maxWidth),
+                                    child: Padding(
+                                      padding: const EdgeInsets.all(8.0),
+                                      child: Text('Alle anderen Orte',
+                                        style: MoreaTextStyle.textField,
+                                      ),
+                                    ),
+                                  ),
+                                  style: ButtonStyle(
+                                      shape: MaterialStateProperty.all<OutlinedBorder>(
+                                          RoundedRectangleBorder(
+                                              borderRadius: BorderRadius.circular(5),
+                                              side: BorderSide(color: Colors.black45))),
+                                      foregroundColor: MaterialStateProperty.all<Color>(
+                                          Colors.black),
+                                      padding: MaterialStateProperty.all<EdgeInsets>(
+                                          EdgeInsets.only(
+                                              top: 10, bottom: 10, left: 5, right: 5)),
+                                      overlayColor: MaterialStateProperty.resolveWith(
+                                              (Set<MaterialState> states) {
+                                            if (states.contains(MaterialState.focused))
+                                              return MoreaColors.violett;
+                                            return null;
+                                          })),
+                                  onPressed: () {
+                                    this.urlMapAbtreten = urlGoogleMapsEmpty;
+                                    this.nameMapAbtreten = 'Alle anderen Orte';
+                                  },
+                                ),
+                              ],
+                            )
+                            /*moreaRaisedButton(this.nameMapAbtreten, () {
                               Navigator.push(
                                   context,
                                   MaterialPageRoute(
@@ -174,16 +276,17 @@ class _ChangeAbtretenState extends State<ChangeAbtreten> {
                                                 47.40548228527181,
                                                 8.559394673386825),
                                             onPlacePicked: (result) {
-                                              this.urlMapAbtreten = result.url;
+                                              this.urlMapAbtreten = result.url!;
                                               this.nameMapAbtreten =
-                                                  result.name;
+                                                  result.name!;
                                               Navigator.of(context).pop();
                                               setState(() {});
                                             },
                                             usePlaceDetailSearch: true,
                                             useCurrentLocation: false,
                                           )));
-                            })),
+                            })*/
+                        ),
                       ],
                     ),
                   ),
@@ -199,7 +302,7 @@ class _ChangeAbtretenState extends State<ChangeAbtreten> {
   Future<Null> _selectTime(BuildContext context) async {
     String hour = zeitAbtreten.split(':')[0];
     String minute = zeitAbtreten.split(':')[1];
-    final TimeOfDay picked = await showTimePicker(
+    final TimeOfDay? picked = await showTimePicker(
       context: context,
       initialTime: TimeOfDay(hour: int.parse(hour), minute: int.parse(minute)),
     );
@@ -230,12 +333,11 @@ class _ChangeAbtretenState extends State<ChangeAbtreten> {
 
   bool saveAndSubmit() {
     final form = _formKey.currentState;
-    if (form.validate()) {
+    if (form!.validate()) {
       if (MoreaInputValidator.url(this.urlMapAbtreten)) {
         form.save();
         return true;
-      }
-      else {
+      } else {
         ScaffoldMessenger.of(context).showSnackBar(SnackBar(
           content: Row(
             children: [
