@@ -1,16 +1,17 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:morea/Widgets/Group/Group_Preview.dart';
 import 'package:morea/services/Group/group_data.dart';
 
 class GroupPreviewButton extends GroupPreview {
-  GroupData groupData;
+  final GroupData groupData;
   @override
-  Widget child;
+  late final Widget child;
   @override
-  double radius;
-  Function function;
-  GroupPreviewButton({@required this.groupData, this.function, this.radius}) {
+  final double radius;
+  final void Function()? function;
+  GroupPreviewButton(
+      {required this.groupData, this.function, required this.radius})
+      : super(groupData: groupData, radius: radius) {
     this.child = InkWell(
       onTap: this.function,
       child: GroupPreview(
