@@ -1,11 +1,10 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:morea/services/morea_firestore.dart';
 
 class GroupRoles extends StatefulWidget {
   final MoreaFirebase moreaFire;
   final String groupID;
-  GroupRoles({@required this.moreaFire, @required this.groupID});
+  GroupRoles({required this.moreaFire, required this.groupID});
 
   GroupRolesState createState() => GroupRolesState();
 }
@@ -16,11 +15,11 @@ class GroupRolesState extends State<GroupRoles> {
     return ListView.builder(
         shrinkWrap: true,
         itemCount:
-            widget.moreaFire.getMapGroupData[widget.groupID].roles.length,
+            widget.moreaFire.getMapGroupData![widget.groupID]!.roles.length,
         itemBuilder: (context, index) {
           return ListTile(
             title: Text(widget
-                .moreaFire.getMapGroupData[widget.groupID].roles.keys
+                .moreaFire.getMapGroupData![widget.groupID]!.roles.keys
                 .toList()[index]),
           );
         });
