@@ -12,25 +12,27 @@ abstract class BaseUrllauncher {
 
   Future<void> openLatLongMaps(double latitude, double longitude);
 
-  Future<void> openLink(String url);
+  Future<void> openLink(Uri url);
 }
 
 class Urllauncher implements BaseUrllauncher {
   Urllauncher();
-  Future<void> openMail(String email, {String subject, String body}) async {
-    String url =
-        'mailto:$email?${subject == null ? "" : ("subject=" + subject)}${body == null ? "" : ("&body=" + body)}';
-    if (await canLaunch(url)) {
-      await launch(url);
+  Future<void> openMail(String email, {String? subject, String? body}) async {
+    Uri url = Uri(scheme: 'mailto', path: email, queryParameters: {
+      'subject': subject == null ? "" : subject,
+      'body': body == null ? '' : body
+    });
+    if (await canLaunchUrl(url)) {
+      await launchUrl(url);
     } else {
       throw 'Could not launch $url';
     }
   }
 
   Future<void> openPhone(String phonenumber) async {
-    String url = 'tel:$phonenumber';
-    if (await canLaunch(url)) {
-      await launch(url);
+    Uri url = Uri(scheme: 'tel', path: phonenumber);
+    if (await canLaunchUrl(url)) {
+      await launchUrl(url);
     } else {
       throw 'Could not launch $url';
     }
@@ -45,18 +47,18 @@ class Urllauncher implements BaseUrllauncher {
   }
 
   Future<void> openLatLongMaps(double latitude, double longitude) async {
-    String googleUrl =
-        'https://www.google.com/maps/search/?api=1&query=$latitude,$longitude';
-    if (await canLaunch(googleUrl)) {
-      await launch(googleUrl);
+    Uri googleUrl = Uri.https(
+        'www.google.com', 'maps/search/?api=1&query=$latitude,$longitude');
+    if (await canLaunchUrl(googleUrl)) {
+      await launchUrl(googleUrl);
     } else {
       throw 'Could not open the map.';
     }
   }
 
-  Future<void> openLink(String url) async {
-    if (await canLaunch(url)) {
-      await launch(url);
+  Future<void> openLink(Uri url) async {
+    if (await canLaunchUrl(url)) {
+      await launchUrl(url);
     } else {
       throw 'Could not launch $url';
     }
