@@ -80,7 +80,7 @@ class _ChangeEmailState extends State<ChangeEmail> {
                           focusedBorder: OutlineInputBorder(borderSide: BorderSide(color: MoreaColors.violett)),
                         ),
                         validator: (value) {
-                          if(value.isEmpty){
+                          if(value!.isEmpty){
                             return 'Bitte nicht leer lassen';
                           } else if (!MoreaInputValidator.email(value)){
                             return 'Bitte gültige E-Mail verwenden';
@@ -102,7 +102,7 @@ class _ChangeEmailState extends State<ChangeEmail> {
 
   bool _validateAndSave() {
     final form = _formKey.currentState;
-    if (form.validate()) {
+    if (form!.validate()) {
       form.save();
       return true;
     } else {
