@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
-Future<Widget> scanDialog(BuildContext context, List<Widget> widgets) {
+Future<dynamic> scanDialog(BuildContext context, List<Widget> widgets) {
   return showDialog(
       context: context,
       builder: (context) {
-        return new Container(
+        return Container(
           color: Colors.black.withOpacity(0.7),
           padding: EdgeInsets.only(top: 40, left: 20, right: 20, bottom: 40),
           child: Card(
