@@ -10,7 +10,7 @@ class DWIFormat implements BaseDWIFormat {
     Iterable<Match> matches = exp.allMatches(str);
     List<String> charakters = ['%'];
     for (Match m in matches) {
-      charakters.add(m.group(0));
+      charakters.add(m.group(0)!);
     }
     charakters.remove('%');
     return charakters.join().toString();
@@ -21,7 +21,7 @@ class DWIFormat implements BaseDWIFormat {
     Iterable<Match> matches = exp.allMatches(str);
     List<String> charakters = ['%'];
     for (Match m in matches) {
-      charakters.add(m.group(0));
+      charakters.add(m.group(0)!);
     }
     charakters.remove('%');
     return charakters.join().toString();
